Update inputsplitter to use new input transformers
